requirements.txt
numpy==1.19.1
dna_features_viewer==3.0.3
matplotlib==3.3.2
gfapy==1.1.0
Bio==0.0.6
Pillow==7.2.0
